+++ tests/utils/utils.js
@@ -10,6 +10,10 @@ export const mockOptionSelection = async (wrapper) => {
   await fireEvent.click(optionElement.element);
 }
 
+export const mockOptionsLoader = async () => {
+  return await Promise.resolve(globalThis.OPTIONS);
+}
+
 export const expandOptionsList = async (wrapper) => {
   const toggle = wrapper.find("#extended__multiselect-toggle");
   await fireEvent.click(toggle.element);
@@ -18,4 +22,4 @@ export const expandOptionsList = async (wrapper) => {
 export const createNewOptionsWrapper = (wrapper) => {
   const optionsWrapper = wrapper.findComponent(ExtendedMultiselectOptions);
   return optionsWrapper.find("div > div:first-child");
-}
+}